Rename `tasks` module to `worker`
`tasks` seems a bit generic, while `worker` makes it a little clearer that this means the background worker
