+++ src/lib.rs
@@ -48,10 +48,10 @@ mod publish_rate_limit;
 pub mod render;
 pub mod schema;
 pub mod sql;
-pub mod tasks;
 mod test_util;
 pub mod uploaders;
 pub mod util;
+pub mod worker;
 
 pub mod controllers;
 pub mod models;
